+++ app/src/main/java/woowacourse/payments/ui/component/NumberTextField.kt
@@ -10,12 +10,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.TextRange
 import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import woowacourse.payments.InputType
 import woowacourse.payments.R
 import woowacourse.payments.ui.theme.Gray100
 import woowacourse.payments.ui.theme.Gray200
