+++ app/src/main/java/woowacourse/payments/ui/transformation/NumberVisualTransformation.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.common
+package woowacourse.payments.ui.transformation
 
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.input.OffsetMapping
